+++ src/pages/index.tsx
@@ -1,14 +1,20 @@
 import React, { useState, useEffect } from "react";
-import Signup from "../components/Signup";
 import { getCourses, getLanguages } from "../lib/backend/database-utils";
-import Course from "../types/Course";
-import Language from "../types/Language";
+import Course from "../types/Course/index";
+import Language from "../types/Language/index";
+import styles from "../styles/Home.module.scss";
+import ReservationForm from "../components/ReservationForm";
+import PageTitle from "../components/PageTitle";
 
 interface HomeProps {
   languages: Language[];
   courses: Course[];
 }
 
+interface SignupProps {
+    languages: Language[];
+    courses: Course[];
+}
 
 export async function getServerSideProps(context: any) {
 
@@ -24,13 +30,20 @@ export async function getServerSideProps(context: any) {
   };
 }
 
-/*
+
+function Signup({ languages, courses }: SignupProps) {
 
 
-TODO: signup should just be this file instead of in a separate component
-
-*/
-
+    return (
+        <div className={styles.container}>
+            <PageTitle title="LT Reservation Form" />
+            <ReservationForm
+                languages={languages}
+                courses={courses}
+            />
+        </div>
+    );
+}
 
 function Home({ languages, courses }: HomeProps) {
